Refactor #165 - move **/view/** to **/ui/**
