+++ app/src/main/res/navigation/nav_home.xml
@@ -21,7 +21,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.nlab.reminder.domain.feature.home.view.HomeFragment"
+        android:name="com.nlab.reminder.domain.feature.home.ui.HomeFragment"
         android:label="Home" />
 
 </navigation>
