+++ build-logic/convention/src/main/kotlin/com/nlab/reminder/convention/AndroidJacoco.kt
@@ -81,6 +81,7 @@ private fun Project.androidJacocoClassDirectories(variant: Variant): Configurabl
                 "**/Manifest*.*",
                 "**/android/**",
                 "**/view/**",
+                "**/ui/**",
                 "**/test/**",
                 "**/di/**",
                 "**/icons/**",
